fix: correct type cast in LogMetrics and date format in SessionPersistence
- LogMetrics: add .toFloat() to avoid Double-to-Float implicit cast
- SessionPersistence: use inline SimpleDateFormat to avoid shared state

+++ app/src/main/java/com/chopcut/ui/components/console/LogMetrics.kt
@@ -100,7 +100,7 @@ class LogMetricsTracker {
         
         val logsPerMinute = if (recentLogs.isNotEmpty()) {
             val duration = (now - recentLogs.first().timestamp) / 1000.0
-            recentLogs.size / duration
+            (recentLogs.size / duration).toFloat()
         } else 0f
         
         val logsPerSecond = lastSecondLogs.size.toFloat()
